[PJRT] Separate collective ops test from TPU runtime test.
